Add action to the ACF append event
When appending fields with ACF the select box is not loaded properly. To fix this issue I propose to add an event listener for the ACF append event. When this event is fired it first removes the existing select box wrapper in the new element and then initialize the select2 plugin for all select boxes.